Refactor project structure by moving UI-related files into a dedicated 'ui' directory and updating include paths
