+++ src/ui/highlighter.cpp
