+++ src/ui/highlighter.h
